requirements.txt
# --- Backend Framework ---
Flask==3.1.2
flask-cors
flask-sqlalchemy
flask-jwt-extended
flask-migrate
pymysql
python-dotenv
python>=3.10,<3.12
Flask-Mail
Flask-APScheduler

# --- Body Posture & Vision ---
mediapipe>=0.10.9
opencv-python
Pillow

# --- Chatbot & RAG ---
google-generativeai
langchain
langchain-community
langchain-google-genai
faiss-cpu
sentence-transformers
beautifulsoup4
lxml==6.0.2
requests

# --- Data Analysis ---
pandas
scikit-learn

# --- Core Dependencies (Versi Aman) ---
pydantic>=2.0
protobuf==4.25.3